(Contrib/lb3361/runjs) latest with direct loading and rom loading
